Updating assignments/java/xor/Xor.java

+++ assignments/java/xor/Xor.java
@@ -93,7 +93,7 @@ public class Xor {
   public String decodeToUtf8(byte[] letterBytes) {
     char[] result = new char[21];
     for (int i = 0; i < letterBytes.length; i++) {
-      result[i] = Character.toChars(letterBytes[i]);
+      result[i] = Character.toChars((int) letterBytes[i]);
       System.out.println(Character.toChars(letterBytes[i]));
     }
     return new String(result);
